♻️ : useFavorite 훅으로 분리
